\name{fEBAcpp-package}
\alias{fEBAcpp-package}
\alias{fEBAcpp}
\docType{package}
\title{
\packageTitle{fEBAcpp}
}
\description{
\packageDescription{fEBAcpp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fEBAcpp}
\packageIndices{fEBAcpp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fEBAcpp}

Maintainer: \packageMaintainer{fEBAcpp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
